fix: add onSuccess, refactor evmContractVerifyForm type
