+++ src/lib/pages/evm-contract-verify/components/solidity/EvmContractVerifySolidityJsonInput.tsx
@@ -1,9 +1,9 @@
 import { Control, useController } from "react-hook-form";
-import { EvmContractVerifyForm } from "../../types";
 import { Heading, Stack } from "@chakra-ui/react";
 import { DropZone } from "lib/components/dropzone";
 import { UploadCard } from "lib/components/upload";
 import { ConstructorArgs } from "../ConstructorArgs";
+import { EvmContractVerifyForm } from "lib/services/types";
 
 interface EvmContractVerifySolidityJsonInputProps {
   control: Control<EvmContractVerifyForm>;
